backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.8.0
attrs==25.3.0
audioread==3.0.1
av==14.2.0
bitarray==3.1.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Cython==3.0.12
decorator==5.2.1
distro==1.9.0
fairseq==0.12.2
faiss-cpu==1.7.3
fakeyou==1.3.0
fastapi==0.115.11
ffmpeg-python==0.2.0
filelock==3.17.0
fish-audio-sdk==2025.2.11
Flask==3.1.0
flask-cors==5.0.1
fpdf==1.7.2
frozenlist==1.5.0
fsspec==2025.3.0
future==1.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-ws==0.7.1
hydra-core==1.0.7
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.43.0
loguru==0.7.3
lxml==5.3.1
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
networkx==3.4.2
numba==0.60.0
numpy==1.23.5
omegaconf==2.0.6
openai==1.66.3
ormsgpack==1.8.0
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
portalocker==3.1.1
praat-parselmouth==0.4.5
propcache==0.3.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-multipart==0.0.20
python-pptx==1.0.2
pywin32==309
pyworld==0.3.5
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
rvc-python==0.1.5
sacrebleu==2.5.1
scikit-learn==1.6.1
scipy==1.15.2
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
starlette==0.46.1
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.5.0
torch==2.6.0
torchaudio==2.6.0
torchcrepe==0.0.23
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
win32_setctime==1.2.0
wsproto==1.2.0
XlsxWriter==3.2.2
yarl==1.18.3
